requirements.txt
fastapi
pydantic
aiofiles
passlib
python-dotenv
